[chore]: rename candygame to carddeck amd models to types
