+++ src/components/Buttons/Button.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
 interface Props {
-    text: string;
+	text: string;
 }
 
 const Button: React.FC<Props> = ({ text }) => {
-    return <button type="button">{text}</button>;
+	return <button type="button">{text}</button>;
 };
 
 export default Button;
