+++ src/components/CardDeck/Title.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 
 interface Props {
-    title: string;
-    subTitle: string;
+	title: string;
+	subTitle: string;
 }
 
 // Component
 const Title: React.FC<Props> = ({ title, subTitle }) => {
-    return (
-        <div>
-            <h1>{title}</h1>
-            <p>{subTitle}</p>
-        </div>
-    );
+	return (
+		<div>
+			<h1>{title}</h1>
+			<p>{subTitle}</p>
+		</div>
+	);
 };
 
 // Export
